Add tests for delete and view commands
Add delete and view command tests for persons and companies and
merged DeleteCompanyCommandParserTest into DeleteCommandParserTest.
