+++ src/test/java/seedu/address/logic/parser/ViewCommandParserTest.java
@@ -3,12 +3,12 @@ package seedu.address.logic.parser;
 import static seedu.address.logic.Messages.MESSAGE_INVALID_COMMAND_FORMAT;
 import static seedu.address.logic.parser.CommandParserTestUtil.assertParseFailure;
 import static seedu.address.logic.parser.CommandParserTestUtil.assertParseSuccess;
+import static seedu.address.testutil.TypicalIndexes.INDEX_FIRST_PERSON_OR_COMPANY;
 
 import org.junit.jupiter.api.Test;
 
-import seedu.address.commons.core.index.Index;
-import seedu.address.logic.commands.deletecommands.DeletePersonCommand;
 import seedu.address.logic.commands.viewcommands.ViewCommand;
+import seedu.address.logic.commands.viewcommands.ViewCompanyCommand;
 import seedu.address.logic.commands.viewcommands.ViewPersonCommand;
 
 public class ViewCommandParserTest {
@@ -16,27 +16,36 @@ public class ViewCommandParserTest {
     private ViewCommandParser parser = new ViewCommandParser();
 
     @Test
-    public void parse_invalidArgs1_throwsParseException() {
-        assertParseFailure(parser, "p a",
-                String.format(MESSAGE_INVALID_COMMAND_FORMAT, DeletePersonCommand.MESSAGE_USAGE));
-    }
-
-    @Test
-    public void parse_invalidArgs2_throwsParseException() {
+    public void parse_invalidArgs_throwsParseException() {
         assertParseFailure(parser, "d 1",
-                String.format(MESSAGE_INVALID_COMMAND_FORMAT, DeletePersonCommand.MESSAGE_USAGE));
-    }
-
-    @Test
-    public void parse_validArgs_returnsViewCommand() {
-        ViewCommand expectedViewCommand =
-                new ViewPersonCommand(Index.fromOneBased(1));
-        assertParseSuccess(parser, "p 1", expectedViewCommand);
+                String.format(MESSAGE_INVALID_COMMAND_FORMAT, ViewCommand.MESSAGE_USAGE));
     }
 
     @Test
     public void parse_notEnoughArgs_throwsParseException() {
-        assertParseFailure(parser, "        ",
-                String.format(MESSAGE_INVALID_COMMAND_FORMAT, DeletePersonCommand.MESSAGE_USAGE));
+        assertParseFailure(parser, "p",
+                String.format(MESSAGE_INVALID_COMMAND_FORMAT, ViewCommand.MESSAGE_USAGE));
+    }
+
+    @Test
+    public void parse_invalidArgsCompany_throwsParseException() {
+        assertParseFailure(parser, "c a",
+                String.format(MESSAGE_INVALID_COMMAND_FORMAT, ViewCompanyCommand.MESSAGE_USAGE));
+    }
+
+    @Test
+    public void parse_invalidArgsPerson_throwsParseException() {
+        assertParseFailure(parser, "p a",
+                String.format(MESSAGE_INVALID_COMMAND_FORMAT, ViewPersonCommand.MESSAGE_USAGE));
+    }
+
+    @Test
+    public void parse_validArgs_returnsDeleteCompanyCommand() {
+        assertParseSuccess(parser, "c 1", new ViewCompanyCommand(INDEX_FIRST_PERSON_OR_COMPANY));
+    }
+
+    @Test
+    public void parse_validArgs_returnsDeletePersonCommand() {
+        assertParseSuccess(parser, "p 1", new ViewPersonCommand(INDEX_FIRST_PERSON_OR_COMPANY));
     }
 }
